recommendation/requirements_backend.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==0.8.11
pydantic==2.12.5
pydantic_core==2.41.5
